requirements.txt
python-dotenv~=0.21.1
SQLAlchemy~=2.0.25
fastapi
asyncpg
uvicorn~=0.15.0
pydantic~=1.10.4
python-jose
nh3
sentry_sdk
uvicorn==0.15.0
jinja2